fix: added detection and marking moved pawns in puzzle preview and edit for coach and student
