+++ src/app/modules/coach/components/coach-puzzle-chessboard-editor/coach-puzzle-chessboard-editor.component.scss
@@ -7,10 +7,15 @@
   aspect-ratio: 1/1;
   display: grid;
   grid-template-columns: repeat(8, 1fr);
+  // max-width: 40rem;
   position: relative;
+  // width: 100%;
+  // width: 40rem;
   width: 32rem;
+  // border: 1px solid red;
   border-radius: 1rem;
   overflow: hidden;
+  // border: 3px solid purple;
 
   .chessboard-rank {
     bottom: 0.25rem;
@@ -86,30 +91,11 @@
 
 .content-container {
   display: flex;
-  width: 60rem;
-  margin: 0 auto;
 
   .header {
     font-size: $font-M;
     font-weight: 700;
     white-space: nowrap;
-  }
-
-  input,
-  textarea {
-    width: 100%;
-    border-radius: 0.5rem;
-    border: 1px solid $border-secondary;
-    font-size: $font-S;
-    padding: 0.5rem;
-    height: 2.6rem;
-    font-family: 'Poppins', sans-serif;
-    font-style: normal;
-  }
-
-  textarea {
-    resize: none;
-    height: calc(1.2rem * 5 + 1rem);
   }
 
   .chessboard-controls {
@@ -120,7 +106,8 @@
     background-color: white;
     border-radius: 1rem;
     min-width: 20rem;
-    padding-top: 20rem;
+    padding-top: 22rem;
+    // border: 3px solid blue;
 
     .operation-btn {
       border: none;
@@ -131,6 +118,7 @@
       width: fit-content;
       padding: 0.5rem 1rem;
       cursor: pointer;
+      // border: 1px solid;
 
       &:hover {
         background-color: color.adjust($primary, $lightness: 40%);
@@ -140,19 +128,27 @@
 
   .chess-content {
     width: 100%;
+    // border: 1px solid red;
+    margin-left: 1rem;
 
     .main-game-content {
       display: flex;
+      // gap: 1rem;
+      // margin-left: calc(50% - 26.5rem);
+      // max-width: 53rem;
       width: 100%;
       gap: 1rem;
+      // border: 3px solid red;
 
       .moves-group {
         flex: 1;
         display: flex;
         flex-direction: column;
+        // justify-content: space-between;
         padding: 1rem;
         background-color: white;
         border-radius: 1rem;
+        // border: 3px solid green;
 
         .nav {
           display: flex;
@@ -188,6 +184,7 @@
           .btn-group {
             display: flex;
             justify-content: space-between;
+            // border: 1px solid;
 
             .square-flex {
               display: flex;
@@ -202,6 +199,7 @@
                 justify-content: center;
                 align-items: center;
                 border-radius: 0.25rem;
+                // border: 1px solid;
               }
 
               .desc {
@@ -211,9 +209,11 @@
           }
 
           .move-list {
+            // height: 100%;
             flex: 1;
             display: flex;
             gap: 0.5rem;
+            // border: 1px solid;
 
             .down-arrow {
               height: inherit;
@@ -247,6 +247,7 @@
       background-color: white;
       border-radius: 1rem;
       margin-top: 1rem;
+      // border: 3px solid pink;
 
       .header {
         margin-bottom: 0.5rem;
