+++ src/app/modules/coach/components/coach-puzzle-chessboard-editor/coach-puzzle-chessboard-editor.component.spec.ts
@@ -1,0 +1,22 @@
+import { ComponentFixture, TestBed } from '@angular/core/testing';
+
+import { PuzzleChessboardComponent } from './coach-puzzle-chessboard-editor.component';
+
+describe('PuzzleChessboardComponent', () => {
+  let component: PuzzleChessboardComponent;
+  let fixture: ComponentFixture<PuzzleChessboardComponent>;
+
+  beforeEach(async () => {
+    await TestBed.configureTestingModule({
+      imports: [PuzzleChessboardComponent]
+    }).compileComponents();
+
+    fixture = TestBed.createComponent(PuzzleChessboardComponent);
+    component = fixture.componentInstance;
+    fixture.detectChanges();
+  });
+
+  it('should create', () => {
+    expect(component).toBeTruthy();
+  });
+});
