+++ src/app/modules/coach/pages/coach-add-puzzle/coach-add-puzzle.component.scss
