--- src/app/modules/coach/pages/coach-exercise-preview/coach-exercise-preview.component.html
@@ -1,21 +1,0 @@
-<div class="chessboard-preview-page">
-  <div class="solution-wrapper">
-    <h2>Solution</h2>
-    <div class="solution-moves">
-      @for (move of solutionArray(); track $index) {
-        @if ($index % 2 === 0) {
-          <div>{{ $index / 2 + 1 }}.</div>
-        }
-        <div class="correct-move">
-          {{ move }}
-        </div>
-      }
-    </div>
-  </div>
-  <div class="chessboard-preview">
-    <app-puzzle-chessboard
-      [startingFen]="fen"
-      [previewMode]="PreviewMode.Coach"
-    ></app-puzzle-chessboard>
-  </div>
-</div>
