--- src/app/modules/coach/pages/coach-exercise-preview/coach-exercise-preview.component.scss
@@ -1,26 +1,0 @@
-.chessboard-preview-page {
-  display: flex;
-  gap: 2rem;
-  justify-content: center;
-}
-
-.chessboard-preview {
-  max-width: 53rem;
-  width: 100%;
-}
-
-.solution-wrapper {
-  margin-top: 7rem;
-}
-
-.solution-moves {
-  display: grid;
-  font-size: 0.875rem;
-  gap: 0.5rem;
-  grid-template-columns: 2.5rem 5rem 5rem;
-  width: 12rem;
-}
-
-.correct-move {
-  font-weight: 700;
-}
