+++ src/app/modules/coach/pages/coach-puzzle-preview/coach-puzzle-preview.component.html
@@ -25,11 +25,11 @@
     <a class="secondary-btn" routerLink="../../../../">Back to course</a>
   </div>
 </div>
-<app-student-puzzle-chessboard
+<app-puzzle-chessboard
   #chessboardRef
   [startingFen]="puzzle.fen"
   [previewMode]="PreviewMode.Student"
   [expectedMoves]="solutionArray()"
   (solved)="showSolvedMessage()"
 >
-</app-student-puzzle-chessboard>
+</app-puzzle-chessboard>
