+++ src/app/modules/coach/pages/coach-puzzle-preview/coach-puzzle-preview.component.ts
@@ -2,11 +2,11 @@ import { Component, computed, inject } from '@angular/core';
 import { ActivatedRoute, RouterLink } from '@angular/router';
 import { DialogModule } from 'primeng/dialog';
 import { PreviewMode } from '../../../shared/enums/preview-mode.enum';
-import { StudentPuzzleChessboardComponent } from '../../../student/components/student-puzzle-chessboard/student-puzzle-chessboard.component';
+import { PuzzleChessboardComponent } from '../../../shared/components/puzzle-chessboard/puzzle-chessboard.component';
 
 @Component({
   selector: 'app-coach-puzzle-preview',
-  imports: [StudentPuzzleChessboardComponent, DialogModule, RouterLink],
+  imports: [PuzzleChessboardComponent, DialogModule, RouterLink],
   templateUrl: './coach-puzzle-preview.component.html',
   styleUrl: './coach-puzzle-preview.component.scss'
 })
