+++ src/app/modules/shared/components/puzzle-chessboard/puzzle-chessboard.component.scss
@@ -7,15 +7,10 @@
   aspect-ratio: 1/1;
   display: grid;
   grid-template-columns: repeat(8, 1fr);
-  // max-width: 40rem;
   position: relative;
-  // width: 100%;
-  // width: 40rem;
   width: 32rem;
-  // border: 1px solid red;
   border-radius: 1rem;
   overflow: hidden;
-  // border: 3px solid purple;
 
   .chessboard-rank {
     bottom: 0.25rem;
@@ -91,11 +86,30 @@
 
 .content-container {
   display: flex;
+  width: 60rem;
+  margin: 0 auto;
 
   .header {
     font-size: $font-M;
     font-weight: 700;
     white-space: nowrap;
+  }
+
+  input,
+  textarea {
+    width: 100%;
+    border-radius: 0.5rem;
+    border: 1px solid $border-secondary;
+    font-size: $font-S;
+    padding: 0.5rem;
+    height: 2.6rem;
+    font-family: 'Poppins', sans-serif;
+    font-style: normal;
+  }
+
+  textarea {
+    resize: none;
+    height: calc(1.2rem * 5 + 1rem);
   }
 
   .chessboard-controls {
@@ -106,8 +120,7 @@
     background-color: white;
     border-radius: 1rem;
     min-width: 20rem;
-    padding-top: 22rem;
-    // border: 3px solid blue;
+    padding-top: 20rem;
 
     .operation-btn {
       border: none;
@@ -118,7 +131,6 @@
       width: fit-content;
       padding: 0.5rem 1rem;
       cursor: pointer;
-      // border: 1px solid;
 
       &:hover {
         background-color: color.adjust($primary, $lightness: 40%);
@@ -128,27 +140,19 @@
 
   .chess-content {
     width: 100%;
-    // border: 1px solid red;
-    margin-left: 1rem;
 
     .main-game-content {
       display: flex;
-      // gap: 1rem;
-      // margin-left: calc(50% - 26.5rem);
-      // max-width: 53rem;
       width: 100%;
       gap: 1rem;
-      // border: 3px solid red;
 
       .moves-group {
         flex: 1;
         display: flex;
         flex-direction: column;
-        // justify-content: space-between;
         padding: 1rem;
         background-color: white;
         border-radius: 1rem;
-        // border: 3px solid green;
 
         .nav {
           display: flex;
@@ -184,7 +188,6 @@
           .btn-group {
             display: flex;
             justify-content: space-between;
-            // border: 1px solid;
 
             .square-flex {
               display: flex;
@@ -199,7 +202,6 @@
                 justify-content: center;
                 align-items: center;
                 border-radius: 0.25rem;
-                // border: 1px solid;
               }
 
               .desc {
@@ -209,11 +211,9 @@
           }
 
           .move-list {
-            // height: 100%;
             flex: 1;
             display: flex;
             gap: 0.5rem;
-            // border: 1px solid;
 
             .down-arrow {
               height: inherit;
@@ -247,7 +247,6 @@
       background-color: white;
       border-radius: 1rem;
       margin-top: 1rem;
-      // border: 3px solid pink;
 
       .header {
         margin-bottom: 0.5rem;
