--- src/app/modules/shared/services/exercise.service.spec.ts
@@ -1,16 +1,0 @@
-import { TestBed } from '@angular/core/testing';
-
-import { ExerciseService } from './exercise.service';
-
-describe('ExerciseService', () => {
-  let service: ExerciseService;
-
-  beforeEach(() => {
-    TestBed.configureTestingModule({});
-    service = TestBed.inject(ExerciseService);
-  });
-
-  it('should be created', () => {
-    expect(service).toBeTruthy();
-  });
-});
