--- src/app/modules/student/models/exercise.model.ts
@@ -1,8 +1,0 @@
-export interface Exercise {
-  id: string;
-  title: string;
-  description: string;
-  fen: string;
-  solution: string;
-  order: number;
-}
