+++ src/app/modules/student/pages/student-solve-puzzle/student-solve-puzzle.component.scss
